requirements.txt
Flask==3.0.1
requests
boto3==1.34.28
mysql-connector-python
python-dotenv
